# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Certificate Manager API",
        "description": "API REST unificada (v1) para la gestión completa de solicitudes de certificados digitales: creación, archivos adjuntos, emisión agnóstica de proveedor (correo legacy / Viafirma RA PKCS#10 Zero-Touch), pagos WOMPI, cuotas POSTPAID, analíticas IA, notificaciones de vencimiento y webhooks. Requiere autenticación OAuth 2.0 con Laravel Passport.",
        "contact": {
            "name": "Soporte Matias",
            "email": "soporte@matias.com.co"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1 — única versión soportada"
        }
    ],
    "paths": {
...
        "/analytics/results": {
            "get": {
                "tags": [
                    "Analíticas IA"
                ],
                "summary": "Listar resultados de análisis IA",
                "description": "Retorna resultados paginados de análisis OCR+IA filtrados por empresa del usuario autenticado.",
                "operationId": "listAnalysisResults",
                "parameters": [
                    {
                        "name": "analysis_type",
                        "in": "query",
                        "description": "Filtrar por tipo de análisis",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "general",
                                "rut",
                                "cedula",
                                "chamber_commerce"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Fecha inicio (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Fecha fin (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de resultados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/DocumentAnalysisResult"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 72
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/analytics/stats": {
            "get": {
                "tags": [
                    "Analíticas IA"
                ],
                "summary": "Estadísticas resumidas de análisis IA",
                "description": "Retorna métricas agregadas: total, completados, fallidos, confianza promedio, tiempo promedio y distribución por tipo.",
                "operationId": "getAnalyticsStats",
                "responses": {
                    "200": {
                        "description": "Estadísticas agregadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AnalyticsStats"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/analytics/results/{id}": {
            "get": {
                "tags": [
                    "Analíticas IA"
                ],
                "summary": "Detalle de un resultado de análisis",
                "description": "Retorna el resultado completo incluyendo texto OCR, respuesta IA, datos extraídos y validación.",
                "operationId": "showAnalysisResult",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del resultado de análisis",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del análisis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DocumentAnalysisResult"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Resultado no encontrado o no pertenece a la empresa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/analytics/providers": {
            "get": {
                "tags": [
                    "Analíticas IA"
                ],
                "summary": "Proveedores IA/OCR activos",
                "description": "Retorna qué proveedores de OCR e IA están configurados y disponibles. Si no hay credenciales, muestra MOCK como proveedor activo.",
                "operationId": "getActiveProviders",
                "responses": {
                    "200": {
                        "description": "Estado de proveedores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProviderStatus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/webhooks/wompi": {
            "post": {
                "tags": [
                    "Pagos Externos"
                ],
                "summary": "Recibir evento de pago de WOMPI",
                "description": "Endpoint receptor de webhooks de WOMPI. La firma HMAC-SHA256 es validada automáticamente por el middleware ValidateWompiSignature. No requiere autenticación Bearer.",
                "operationId": "c8d42e7b0b7081379abe026bee89c703",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "event": {
                                        "type": "string",
                                        "example": "transaction.updated"
                                    },
                                    "data": {
                                        "properties": {
                                            "transaction": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "reference": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "PENDING",
                                                            "APPROVED",
                                                            "DECLINED",
                                                            "VOIDED",
                                                            "ERROR"
                                                        ]
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "timestamp": {
                                        "type": "integer",
                                        "example": 1745000000
                                    },
                                    "signature": {
                                        "properties": {
                                            "checksum": {
                                                "type": "string"
                                            },
                                            "properties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento recibido y encolado"
                    },
                    "401": {
                        "description": "Firma HMAC inválida"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DocumentAnalysisResult": {
                "description": "Resultado de análisis IA de un documento (OCR + IA)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "certificate_request_id": {
                        "type": "integer",
                        "example": 42
                    },
                    "file_manager_id": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "provider": {
                        "description": "Proveedor IA",
                        "type": "string",
                        "example": "GEMINI"
                    },
                    "analysis_type": {
                        "type": "string",
                        "enum": [
                            "general",
                            "rut",
                            "cedula",
                            "chamber_commerce"
                        ],
                        "example": "rut"
                    },
                    "ocr_text": {
                        "description": "Texto extraído por OCR",
                        "type": "string",
                        "nullable": true
                    },
                    "ocr_provider": {
                        "description": "Proveedor OCR",
                        "type": "string",
                        "example": "GOOGLE_VISION",
                        "nullable": true
                    },
                    "ocr_confidence": {
                        "type": "number",
                        "format": "float",
                        "example": 0.87,
                        "nullable": true
                    },
                    "ai_response": {
                        "description": "Respuesta estructurada del análisis IA",
                        "type": "object",
                        "nullable": true
                    },
                    "ai_confidence": {
                        "type": "number",
                        "format": "float",
                        "example": 0.92,
                        "nullable": true
                    },
                    "completeness_score": {
                        "type": "number",
                        "format": "float",
                        "example": 0.85,
                        "nullable": true
                    },
                    "extracted_data": {
                        "description": "Datos clave extraídos del documento",
                        "type": "object",
                        "nullable": true
                    },
                    "validation_result": {
                        "description": "Resultado de validación",
                        "type": "object",
                        "nullable": true
                    },
                    "processing_time": {
                        "description": "Tiempo en segundos",
                        "type": "number",
                        "format": "float",
                        "example": 2.345,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "PROCESSING",
                            "COMPLETED",
                            "FAILED"
                        ],
                        "example": "COMPLETED"
                    },
                    "error_message": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "AnalyticsStats": {
                "description": "Estadísticas agregadas del módulo de análisis IA",
                "properties": {
                    "total": {
                        "type": "integer",
                        "example": 150
                    },
                    "completed": {
                        "type": "integer",
                        "example": 142
                    },
                    "failed": {
                        "type": "integer",
                        "example": 8
                    },
                    "avg_confidence": {
                        "type": "number",
                        "format": "float",
                        "example": 0.89
                    },
                    "avg_processing_time": {
                        "description": "Promedio en segundos",
                        "type": "number",
                        "format": "float",
                        "example": 1.876
                    },
                    "avg_completeness": {
                        "type": "number",
                        "format": "float",
                        "example": 0.82
                    },
                    "by_type": {
                        "description": "Conteo por tipo de análisis",
                        "properties": {
                            "general": {
                                "type": "integer",
                                "example": 50
                            },
                            "rut": {
                                "type": "integer",
                                "example": 40
                            },
                            "cedula": {
                                "type": "integer",
                                "example": 35
                            },
                            "chamber_commerce": {
                                "type": "integer",
                                "example": 25
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ProviderStatus": {
                "description": "Estado de los proveedores IA/OCR activos",
                "properties": {
                    "ocr": {
                        "properties": {
                            "provider": {
                                "type": "string",
                                "example": "GOOGLE_VISION"
                            },
                            "available": {
                                "type": "boolean",
                                "example": true
                            }
                        },
                        "type": "object"
                    },
                    "ai": {
                        "properties": {
                            "provider": {
                                "type": "string",
                                "example": "GEMINI"
                            },
                            "available": {
                                "type": "boolean",
                                "example": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de acceso OAuth 2.0 obtenido desde /authentication/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}